requirements.txt
streamlit
python-docx
fpdf
openai
Pillow
